Push Five
Added
additional_info functionality
address  functionality
creditcard add/update/delete/getcardbyemail
neighbourhood_info
login and signup
rewards

+++ src/main/java/com/example/rems/domain/BookingDetails.java
@@ -23,7 +23,7 @@ public class BookingDetails {
 
     private BigInteger creditcard_number;
 
-    private BigInteger booking_amount;
+    private Integer booking_amount;
 
     private Date start_date;
